requirements.txt
amqp==5.3.1
ansible-core==2.14.17
anyio==4.9.0
argcomplete==1.12.0
asgiref==3.8.1
async-timeout==5.0.1
backoff==2.2.1
billiard==4.2.1
blinker==1.9.0
blivet==3.6.0
Brlapi==0.8.2
celery==5.5.3
cffi==1.14.5
chardet==4.0.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cockpit @ file:///builddir/build/BUILD/cockpit-323.1/tmp/wheel/cockpit-323.1-py3-none-any.whl
coverage==6.2
cron-descriptor==1.4.5
cryptography==36.0.1
cupshelpers==1.0
dasbus==1.4
dbus-python==1.2.18
decorator==4.4.2
distro==1.5.0
dj-database-url==3.0.1
Django==4.2.21
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-crontab==0.7.1
django-environ==0.12.0
django-filter==25.1
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
drf-yasg==1.21.10
et_xmlfile==2.0.0
exceptiongroup==1.3.0
file-magic==0.4.0
Flask==3.1.0
gpg==1.15.1
gql==3.5.3
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==2.10
importlib_metadata==8.6.1
inflection==0.5.1
iniparse==0.4
initial-setup==0.3.90.2
itsdangerous==2.2.0
jeepney==0.9.0
Jinja2==3.1.6
kombu==5.5.4
langtable==0.0.54
libcomps==0.1.18
lxml==4.6.5
MarkupSafe==3.0.2
multidict==6.6.3
nftables==0.1
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
parameterized==0.9.0
perf==0.1
pexpect==4.8.0
pid==2.2.3
ply==3.11
productmd==1.31
promise==2.3
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==5.8.0
psycopg2-binary==2.9.10
ptyprocess==0.6.0
pwquality==1.4.4
pycairo==1.20.1
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
pyenchant==3.2.0
PyGObject==3.40.1
pyinotify==0.9.6
PyJWT==2.9.0
pykickstart==3.32.11
pyotp==2.9.0
pyparsing==2.4.7
pyparted==3.12.0
PyQt5==5.15.9
PyQt5-sip==12.11.1
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.1.1
python-linux-procfs==0.7.3
python-meh==0.50
pytz==2025.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
redis==6.3.0
rednotebook==2.29.6
requests==2.25.1
requests-file==1.5.1
requests-ftp==0.3.1
requests-toolbelt==1.0.0
resolvelib==0.5.4
rpm==4.16.1.3
SecretStorage==3.3.3
selinux==3.6
sepolicy==3.6
setools==4.4.4
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.32/src
simpleline==1.8.3
six==1.15.0
sniffio==1.3.1
sos==4.8.2
sqlparse==0.5.3
subscription-manager==1.29.42
systemd-python==234
text-unidecode==1.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.5
vboxapi==1.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.20.1
zipp==3.21.0
